backend/foodgram/requirements.txt
asgiref==3.5.2
Django==3.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.4.0
sqlparse==0.4.3
flake8
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2020.1
djoser
Pillow
psycopg2-binary==2.8.6
reportlab